backend/requirements.txt
selenium==4.25.0
bs4
matplotlib_scalebar
contextily
rasterstats
tqdm
statsmodels>=0.13.0
seaborn
pymannkendall
scikit-image
opencv-python
affine==2.4.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.1
coverage==7.6.11
cryptography==44.0.2
cycler==0.12.1
decorator==5.1.1
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-extensions==3.2.3
django-leaflet==0.31.0
djangorestframework==3.15.2
ecdsa==0.19.1
executing==2.2.0
factory_boy==3.3.3
Faker==35.2.0
flake8==7.1.1
fonttools==4.56.0
GeoAlchemy2==0.17.1
geojson==3.2.0
geopandas==1.0.1
greenlet==3.1.1
gsconfig-py3==1.0.8

h11==0.14.0
idna==3.10
inflect==7.5.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.32.0
jedi==0.19.2
joblib==1.4.2
kiwisolver==1.4.8
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
more-itertools==10.6.0
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
rasterio==1.4.3
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
sentry-sdk==2.20.0
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3

threadpoolctl==3.5.0
traitlets==5.14.3
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0

wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
pykrige
fiona

reportlab

